+++ nautobot_secrets_providers/providers/__init__.py
@@ -1,14 +1,13 @@
 """Nautobot Secrets Providers."""
 
-from .aws import AWSSecretsManagerSecretsProvider
-from .aws_ssm import AWSSystemsManagerParameterStore
+from .aws import AWSSecretsManagerSecretsProvider, AWSSystemsManagerParameterStore
 from .hashicorp import HashiCorpVaultSecretsProvider
 from .delinea import ThycoticSecretServerSecretsProviderId, ThycoticSecretServerSecretsProviderPath
 
-__all__ = (  # type: ignore
-    AWSSecretsManagerSecretsProvider,  # pylint: disable=invalid-all-object
-    HashiCorpVaultSecretsProvider,  # pylint: disable=invalid-all-object
-    AWSSystemsManagerParameterStore,  # pylint: disable=invalid-all-object
-    ThycoticSecretServerSecretsProviderId,  # pylint: disable=invalid-all-object
-    ThycoticSecretServerSecretsProviderPath,  # pylint: disable=invalid-all-object
+__all__ = (
+    "AWSSecretsManagerSecretsProvider",
+    "HashiCorpVaultSecretsProvider",
+    "AWSSystemsManagerParameterStore",
+    "ThycoticSecretServerSecretsProviderId",
+    "ThycoticSecretServerSecretsProviderPath",
 )
